requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.7
bcrypt==4.0.1
flask-security-too
python-dotenv
flask-sqlalchemy
flask-cors==4.0.0
flask-migrate
marshmallow